requirements.txt
Flask==0.12
matplotlib==1.5.3
numpy==1.11.3
pyzmq==16.0.2
scikit-learn==0.18.1
Scrapy==1.3.0
